pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cyphera-keychain"
version = "0.1.0"
description = "Key provider abstraction for Cyphera encryption SDKs"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Cyphera", email = "sdk@cyphera.dev" }]
keywords = ["cyphera", "encryption", "key-management", "fpe"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
]

[project.urls]
Homepage = "https://github.com/cyphera-labs/keychain-python"

[project.optional-dependencies]
aws = ["boto3>=1.34"]
gcp = ["google-cloud-kms>=2.21"]
azure = ["azure-keyvault-keys>=4.9", "azure-identity>=1.16"]
vault = ["hvac>=2.1"]
dev = ["pytest>=8", "botocore>=1.34"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests requiring live cloud emulators",
]
